LibWinDog/Platforms/Mastodon/requirements.txt
Mastodon.py
beautifulsoup4
